requirements.txt
numpy
pandas
tensorflow==2.12
streamlit
seaborn
matplotlib
ensure
pyYAML
tqdm
py7zr
datasets
boto3
mypy-boto3-s3
ensure==1.0.2
fastapi== 0.78.0
Jinja2==3.1.2
uvicorn == 0.18.3
python-box==6.0.2
notebook